RUE-859: split call and intrinsic analysis
