+++ crates/rue-air/src/sema/analysis.rs
@@ -2517,83 +2517,6 @@ where
     Ok(())
 }
 
-/// Shared checks for a module-member function call. Validates module
-/// membership, visibility, and arity against the callee.
-///
-/// Membership (spec 4.13:90, RUE-140): a module contains only declarations
-/// from the imported file. The callee is resolved by the receiver module's
-/// canonical `FileId`, and `member_file_id` is checked as a defensive invariant.
-/// Comparing by canonical FileId (rather
-/// than raw path string) makes equivalent import spellings — `helper.rue` vs
-/// `./helper.rue` — resolve members identically (spec 10.2:4, RUE-240).
-///
-/// Two operations deliberately stay per-pipeline at the call sites: argument
-/// analysis (it recurses into the owning pipeline) and the exclusive-access
-/// check (whose shared core is `check_exclusive_access_in`, RUE-141).
-#[allow(clippy::too_many_arguments)]
-fn check_module_member_call(
-    module_name: &str,
-    module_file_id: Option<FileId>,
-    member_file_id: FileId,
-    fn_name_str: &str,
-    param_types: &[Type],
-    argument_count: usize,
-    accessible: bool,
-    via_reexport: bool,
-    span: Span,
-) -> CompileResult<()> {
-    // Check membership: the function must be defined in the module's file
-    // (canonical FileId equality) — UNLESS the call resolved through a
-    // re-export const in the facade, whose presence is the membership grant
-    // (`pub const f = @import("x").f;`, ADR-0026, RUE-592).
-    if !via_reexport && module_file_id != Some(member_file_id) {
-        return Err(CompileError::new(
-            ErrorKind::UnknownModuleMember {
-                module_name: module_name.to_string(),
-                member_name: fn_name_str.to_string(),
-            },
-            span,
-        ));
-    }
-
-    // Check visibility: private functions are only accessible from the same directory
-    if !accessible {
-        return Err(CompileError::new(
-            ErrorKind::PrivateMemberAccess {
-                item_kind: "function".to_string(),
-                name: fn_name_str.to_string(),
-            },
-            span,
-        ));
-    }
-
-    // Check argument count
-    if argument_count != param_types.len() {
-        return Err(CompileError::new(
-            ErrorKind::WrongArgumentCount {
-                expected: param_types.len(),
-                found: argument_count,
-            },
-            span,
-        ));
-    }
-
-    Ok(())
-}
-
-/// Encode the analyzed arguments and emit the Call instruction for a
-/// module-member function call.
-fn emit_module_member_call(
-    air: &mut Air,
-    function_name: Spur,
-    air_args: &[AirCallArg],
-    return_type: Type,
-    span: Span,
-) -> CompileResult<AnalysisResult> {
-    let air_ref = air.add_call(None, function_name, air_args, return_type, span)?;
-    Ok(AnalysisResult::new(air_ref, return_type))
-}
-
 impl<'a> BodySema<'a> {
     /// Create a type mismatch error with safe type name resolution.
     ///
